requirements.txt
streamlit
openai
pydantic
google-generativeai
python-dotenv